requirements.txt
discord.py
jishaku
dotenv